single file component: StartScreen (#3408)
